World_separation
Created World.cpp and .h
Tested for functionality
